types: extract shared CoS lease cluster into types/shared_cos_lease.rs
#1035 P4 — convert types.rs → types/mod.rs (directory module) and
extract the shared CoS lease + MQFQ V_min coordination cluster
(structures + impls + their inline test mod) into a sibling file.

Surgically extracted only the lease cluster (lines 1421-2067 of
the old types.rs):
- struct SharedCoSQueueLease + impl
- struct SharedCoSRootLease + impl
- struct PaddedVtimeSlot + impl + const NOT_PARTICIPATING
- struct SharedCoSQueueVtimeFloor + impl
- struct SharedCoSLeaseConfig (private)
- struct SharedCoSLeaseState (private)
- consts COS_ROOT_LEASE_TARGET_US / MIN_BYTES / MAX_BYTES
- the inline #[cfg(test)] mod tests block that exercises them

The boundary stops BEFORE the unrelated CoSPendingTxItem /
COS_PRIORITY_LEVELS / WorkerCommand / etc. types that follow at
line 2069+ in the original file. (Earlier P1 attempt grabbed too
much because it used a doc-comment as boundary.)

types/mod.rs production LOC (excluding mod tests):
  2,009 → **1,580** — under the 2,000 modularity-discipline
  threshold for the first time since this file grew over.
types/shared_cos_lease.rs: ~650 LOC including tests.

Visibility: items widened from pub(super) to pub(in crate::afxdp)
so the existing external paths (super::types::SharedCoS*,
super::types::PaddedVtimeSlot, etc.) keep resolving from afxdp/*.

build clean, 865 Rust tests pass.
